move project updates load down to root of dom
