+++ src/components/ProjectDashboard/ProjectDashboard.tsx
@@ -12,6 +12,7 @@ import { PayProjectCard } from './components/PayProjectCard'
 import { ProjectCartProvider } from './components/ProjectCartProvider'
 import { ProjectHeader } from './components/ProjectHeader'
 import { ProjectTabs } from './components/ProjectTabs'
+import { ProjectUpdatesProvider } from './components/ProjectUpdatesProvider'
 import { SuccessPayView } from './components/SuccessPayView'
 import { useProjectPageQueries } from './hooks/useProjectPageQueries'
 
@@ -26,33 +27,35 @@ export const ProjectDashboard = () => {
       <AnnouncementLauncher>
         <FundingCycleCountdownProvider>
           <ProjectCartProvider>
-            <div className="flex w-full flex-col items-center pb-48">
-              {projectPayReceipt !== undefined ? (
-                <SuccessPayView />
-              ) : (
-                <>
-                  <CoverPhoto />
-                  <div className="flex w-full justify-center md:px-6">
-                    <div className="flex w-full max-w-6xl flex-col">
-                      <ProjectHeader className="mt-4 px-4 md:px-0" />
-                      <div className="mt-10 flex w-full flex-col gap-4 px-4 md:flex-row md:px-0">
-                        <PayProjectCard className="flex-1" />
-                        {shouldShowNftCard ? <NftRewardsCard /> : null}
-                        <CurrentCycleCard
-                          className={twMerge(
-                            'hidden w-full md:max-w-sm',
-                            shouldShowNftCard ? 'lg:block' : 'md:block',
-                          )}
-                        />
+            <ProjectUpdatesProvider>
+              <div className="flex w-full flex-col items-center pb-48">
+                {projectPayReceipt !== undefined ? (
+                  <SuccessPayView />
+                ) : (
+                  <>
+                    <CoverPhoto />
+                    <div className="flex w-full justify-center md:px-6">
+                      <div className="flex w-full max-w-6xl flex-col">
+                        <ProjectHeader className="mt-4 px-4 md:px-0" />
+                        <div className="mt-10 flex w-full flex-col gap-4 px-4 md:flex-row md:px-0">
+                          <PayProjectCard className="flex-1" />
+                          {shouldShowNftCard ? <NftRewardsCard /> : null}
+                          <CurrentCycleCard
+                            className={twMerge(
+                              'hidden w-full md:max-w-sm',
+                              shouldShowNftCard ? 'lg:block' : 'md:block',
+                            )}
+                          />
+                        </div>
+                        <ProjectTabs className="mt-8" />
                       </div>
-                      <ProjectTabs className="mt-8" />
                     </div>
-                  </div>
-                </>
-              )}
-            </div>
-            <Footer />
-            <Cart />
+                  </>
+                )}
+              </div>
+              <Footer />
+              <Cart />
+            </ProjectUpdatesProvider>
           </ProjectCartProvider>
         </FundingCycleCountdownProvider>
       </AnnouncementLauncher>
