+++ src/components/ProjectDashboard/components/ProjectUpdatesProvider/index.ts
@@ -1,0 +1,1 @@
+export * from './ProjectUpdatesProvider'
